Aerial Object Classification & Detection/requirements.txt
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
pillow>=10.0.0
opencv-python>=4.8.0
tqdm>=4.66.0
pytest>=8.0.0